Validate layer settings before setting them
This nullified the need for rawSettings vs settings, so they were merged
in favor of using settingsDescription for en-/de-coding.

I hate these big commits where a relatively small change in design
affects all of the layers and requires changes in all of them. I need to
figure out how to refactor without that. Maybe allow both the old and
the new way until things are figured out?